Proof the tut's outline
Good job! You're getting clearer and more precise already.
It looks like I rewrote everything because I took the opportunity to:

1. Reorder some lines. When you introduce a concept or a technique you want to
describe it and show its effect/explain why we can use it first, then move on to
the steps to learn the tool
2. Proof the english. For general pointers check
http://docs.godotengine.org/en/latest/community/contributing/docs_writing_guidelines.html
that we wrote for the Godot docs. Favoring the direct voice and action verbs
helps shorten sentences and convey information more precisely
3. Add a handful of small notes from my experience working in art and doing UX design.